# docs/api/HANA-COMMERCE-OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Henna internal pilot commerce API",
    "version": "2026-10-04",
    "description": "Backend contract only; SMS, PSP and logistics excluded. Command and stored resource records retain PascalCase; public typed records use camelCase. Runtime OpenAPI can remain generic for JsonElement endpoints."
  },
  "paths": {
...
    "/api/v1/carts/current/items": {
      "post": {
        "description": "Authenticated owner; alias of SET_CART_ITEM. A supplied body ID must match the route ID.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "quantity": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "maximum": 999
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "maximum": 2147483647
                  }
                },
                "required": [
                  "productId",
                  "quantity"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes": {
      "post": {
        "description": "Authenticated owner; alias of CREATE_QUOTE. A supplied body ID must match the route ID.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sellerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "addressId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "purchaseType": {
                    "type": "string",
                    "enum": [
                      "PERSONAL",
                      "LEGAL"
                    ]
                  },
                  "fulfillmentMode": {
                    "type": "string",
                    "enum": [
                      "PICKUP"
                    ]
                  }
                },
                "required": [
                  "sellerId",
                  "addressId",
                  "purchaseType",
                  "fulfillmentMode"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "post": {
        "description": "Authenticated owner; alias of PLACE_ORDER. A supplied body ID must match the route ID.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quoteId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "creditGrantId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "unavailableDisposition": {
                    "type": "string",
                    "enum": [
                      "KEEP",
                      "REMOVE"
                    ]
                  },
                  "confirmUnavailable": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "quoteId",
                  "unavailableDisposition"
                ]
              }
            }
          }
        }
      },
      "get": {
        "description": "Scoped before pagination. /me and /orders show the buyer; /seller requires active SELLER.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "format": "int64",
                      "minimum": 1
                    },
                    "pageSize": {
                      "const": 20
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        }
      }
    },
    "/api/v1/me/addresses": {
      "post": {
        "description": "Authenticated owner; alias of SAVE_ADDRESS. A supplied body ID must match the route ID.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerAddress"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "addressId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "cityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  }
                },
                "required": [
                  "addressId",
                  "cityId",
                  "text",
                  "latitude",
                  "longitude"
                ]
              }
            }
          }
        }
      },
      "get": {
        "description": "Scoped before pagination. /me and /orders show the buyer; /seller requires active SELLER.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BuyerAddress"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "format": "int64",
                      "minimum": 1
                    },
                    "pageSize": {
                      "const": 20
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        }
      }
    },
    "/api/v1/seller/offers": {
      "post": {
        "description": "Active SELLER; alias of SAVE_OFFER. A supplied body ID must match the route ID.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Offer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "offerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "priceRial": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 1
                  },
                  "stock": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "maximum": 2147483647
                  }
                },
                "required": [
                  "offerId",
                  "productId",
                  "priceRial",
                  "stock",
                  "expectedVersion"
                ]
              }
            }
          }
        }
      },
      "get": {
        "description": "Scoped before pagination. /me and /orders show the buyer; /seller requires active SELLER.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Offer"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "format": "int64",
                      "minimum": 1
                    },
                    "pageSize": {
                      "const": 20
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        }
      }
    },
    "/api/v1/me/withdrawals": {
      "post": {
        "description": "Authenticated owner; alias of REQUEST_WITHDRAWAL. A supplied body ID must match the route ID.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Withdrawal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountRial": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 1
                  },
                  "ibanVerificationRequestReference": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 240
                  }
                },
                "required": [
                  "amountRial",
                  "ibanVerificationRequestReference"
                ]
              }
            }
          }
        }
      },
      "get": {
        "description": "Scoped before pagination. /me and /orders show the buyer; /seller requires active SELLER.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 10000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Withdrawal"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "format": "int64",
                      "minimum": 1
                    },
                    "pageSize": {
                      "const": 20
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        }
      }
    },
...
    "/api/v1/orders/{id}": {
      "get": {
        "description": "Buyer, current seller or authorized support/admin; other users receive 404.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or missing UUID Idempotency-Key"
          },
          "401": {
            "description": "Missing/invalid/expired session"
          },
          "403": {
            "description": "Current role or ownership denied"
          },
          "404": {
            "description": "Missing or concealed resource"
          },
          "409": {
            "description": "State, balance, inventory, version or idempotency conflict"
          },
          "429": {
            "description": "New command limit exceeded"
          },
          "503": {
            "description": "Backend/integration unconfigured or temporarily unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "Current server-backed opaque session; not a user-provided account ID."
      }
    },
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "BuyerId": {
            "type": "string",
            "format": "uuid"
          },
          "Items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "Version": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "Id",
          "BuyerId",
          "Items",
          "Version"
        ]
      },
...
      "BuyerAddress": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "BuyerId": {
            "type": "string",
            "format": "uuid"
          },
          "CityId": {
            "type": "string",
            "format": "uuid"
          },
          "Text": {
            "type": "string"
          },
          "Latitude": {
            "type": "number"
          },
          "Longitude": {
            "type": "number"
          }
        },
        "required": [
          "Id",
          "BuyerId",
          "CityId",
          "Text",
          "Latitude",
          "Longitude"
        ]
      },
...
    }
  }
}